+++ library/src/getitune/backend/openvino/models/instance_segmentation.py
@@ -152,7 +152,10 @@ class OVInstanceSegmentationModel(OVModel):
                     dtype=torch.float32,
                 ),
             )
-            scores.append(torch.tensor(output.scores.reshape(-1)))
+            # Cast scores to float32 to match the float32 bboxes. ModelAPI returns numpy
+            # arrays whose default float dtype is float64; a float64/float32 mismatch makes
+            # torchvision NMS fail ("dets should have the same type as scores") during tile merge.
+            scores.append(torch.tensor(output.scores.reshape(-1), dtype=torch.float32))
 
             raw_masks = torch.tensor(output.masks)
             if raw_masks.shape[-2:] == (ori_h, ori_w):
